Ignore the contents of the `tests/files/fs` dir
This directory can be used to test the file system in cases where the
file system virtualization methods cannot be used (such as when certain
file system functions are not available to custom stream wrappers). This
directory is cleared or created at the beginning of *all* tests in the
FileTest test.